HDDS-16076. Separate test class for testInstallSnapshotDuringBootstrapping.
